stream_sed and stream_list fix for arcgis provider #81.
If the stream has n reaches:
In arcgis, stream_seq matrix values are in range 1..n, but streamlist
dicrionary FID values are in range 0..(n-1). There fore ther is new -1
in smoderp2d/providers/arcgis/stream_preparation.py.
In grass (if I understand) stream_seq matrix and streamlist dictionary
valuse area the some values, for instance 1,3,5.
For grass provider the function _get_mat_stream_seg_ is the same as the one
deleted from base provider.

This commit solves error commented in comment
https://github.com/storm-fsv-cvut/smoderp2d/pull/82#issuecomment-622002070

 On branch issue_81_stream_fid
 Your branch is up to date with 'origin/issue_81_stream_fid'.

 Changes to be committed:
	modified:   smoderp2d/providers/arcgis/stream_preparation.py
	modified:   smoderp2d/providers/base/stream_preparation.py
	modified:   smoderp2d/providers/grass/stream_preparation.py
