+++ smoderp2d/providers/grass/stream_preparation.py
@@ -9,6 +9,7 @@ from grass.pygrass.vector import Vector, VectorTopo
 from smoderp2d.providers.base.stream_preparation import StreamPreparationBase
 from smoderp2d.providers.base.stream_preparation import StreamPreparationError, ZeroSlopeError
 from smoderp2d.providers.grass.manage_fields import ManageFields
+from smoderp2d.core.general import Globals as Gl
 
 class StreamPreparation(StreamPreparationBase, ManageFields):
     def __init__(self, args, writter):
@@ -238,6 +239,15 @@ class StreamPreparation(StreamPreparationBase, ManageFields):
         self._get_mat_stream_seg_(mat_stream_seg, no_of_streams)
 
         return mat_stream_seg
+
+    def _get_mat_stream_seg_(self, mat_stream_seg, no_of_streams):
+        # each element of stream has a number assigned from 0 to
+        # no. of stream parts
+        for i in range(self.rows):
+            for j in range(self.cols):
+                if mat_stream_seg[i][j] > 0: # FID starts at 1
+                    # state 0|1|2 (> Gl.streams_flow_inc -> stream flow)
+                    mat_stream_seg[i][j] += Gl.streams_flow_inc
 
     def _stream_slope(self, stream):
         """
